Sample test for path parser
